requirements.txt
aws-cdk.core==1.32.1
aws-cdk.aws-ec2==1.32.1
aws-cdk.aws-ecr==1.32.1
aws-cdk.aws-ecs==1.32.1
aws-cdk.aws-ecs-patterns==1.32.1
aws-cdk.aws-rds==1.32.1